// clubhub-backend_final/clubhub-backend_final/src/auth/dtos/update-user.dto.ts
// update-user.dto.ts

export class UpdateUserDto {
    name?: string;  // Optional fields for updating
    email?: string;
    idCard?: string;
    role?: string;
    phone?: string;
    address?: string;
    academicGrade?: string;
    linkedin?: string;
    instagram?: string;
    profileImage?: string;
  }
  